PERSISTENT-DATAFRAME: Rename module to fit the class name
